requirements.txt
requests
kafka-python
pyspark
cassandra-driver